feat(catalog): support adding location analyzers in processing extension
