+++ .changeset/blue-bags-warn.md
@@ -1,0 +1,7 @@
+---
+'@backstage/plugin-catalog-backend': minor
+'@backstage/plugin-catalog-node': minor
+'@backstage/plugin-catalog-backend-module-github': patch
+---
+
+Support adding location analyzers in catalog processing extension point and move `AnalyzeOptions` and `ScmLocationAnalyzer` types to `@backstage/plugin-catalog-node`
