+++ plugins/catalog-backend-module-github/api-report.md
@@ -3,7 +3,7 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-import { AnalyzeOptions } from '@backstage/plugin-catalog-backend';
+import { AnalyzeOptions } from '@backstage/plugin-catalog-node';
 import { CatalogProcessor } from '@backstage/plugin-catalog-node';
 import { CatalogProcessorEmit } from '@backstage/plugin-catalog-node';
 import { Config } from '@backstage/config';
@@ -21,7 +21,7 @@ import { Logger } from 'winston';
 import { PluginEndpointDiscovery } from '@backstage/backend-common';
 import { PluginTaskScheduler } from '@backstage/backend-tasks';
 import { ScmIntegrationRegistry } from '@backstage/integration';
-import { ScmLocationAnalyzer } from '@backstage/plugin-catalog-backend';
+import { ScmLocationAnalyzer } from '@backstage/plugin-catalog-node';
 import { TaskRunner } from '@backstage/backend-tasks';
 import { TokenManager } from '@backstage/backend-common';
 import { UserEntity } from '@backstage/catalog-model';
