+++ plugins/catalog-backend/src/ingestion/index.ts
@@ -21,6 +21,12 @@ export type {
   AnalyzeLocationRequest,
   AnalyzeLocationResponse,
   LocationAnalyzer,
-  ScmLocationAnalyzer,
-  AnalyzeOptions,
 } from './types';
+
+/**
+ * @deprecated use the same type from `@backstage/plugin-catalog-node` instead
+ */
+export type {
+  AnalyzeOptions,
+  ScmLocationAnalyzer,
+} from '@backstage/plugin-catalog-node';
