+++ plugins/catalog-node/src/processing/index.ts
@@ -15,9 +15,11 @@
  */
 
 export type {
+  AnalyzeOptions,
   DeferredEntity,
   PlaceholderResolver,
   PlaceholderResolverParams,
   PlaceholderResolverRead,
   PlaceholderResolverResolveUrl,
+  ScmLocationAnalyzer,
 } from './types';
